Move notifications definitions to its own file
